2025-T2/document-parser/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pdfminer.six>=20231228
python-docx>=1.1.0
nltk>=3.9.1
PyPDF2>=3.0.1
